requirements.txt
# Core
PyYAML>=6.0
Pillow>=10.0.0
numpy>=1.26,<2.4
scikit-image>=0.22.0

# Deep learning
torch>=2.7.0
torchvision>=0.22.0
transformers>=5.0.0
timm>=1.0.17

# SAM 3 (Meta Segment Anything Model 3)
sam3>=0.1.3
psutil

# HuggingFace model hub (for checkpoint downloads)
huggingface-hub>=1.1.4
